Refactor: Decouple handle.cpp and standardize project structure.
- Extracted `async_process.h` and `request.h` to reduce coupling in
`handle.cpp`.
- Unified comment style across the project for consistency.
- Introduced namespaces `ytweb`.
