+++ src/handle.h
@@ -2,7 +2,12 @@
 
 #include "webui.hpp"
 
+namespace ytweb
+{
+
 void handle_submit_url(webui::window::event* event);
 
-// Send a log message to the frontend
+// Send a log message to the frontend.
 void send_log(webui::window::event* event, std::string const& message);
+
+}  // namespace ytweb
