+++ src/main.cpp
@@ -11,7 +11,7 @@ fs::path const INDEX_PATH = YT_DLP_WEB_PATH;
 
 int main(int argc, char* argv[])
 {
-    // Check if the index.html file exists
+    // Check if the index.html file exists.
     if (!fs::exists(INDEX_PATH) || !fs::is_directory(INDEX_PATH))
     {
         std::cerr << "Path not exist: " << INDEX_PATH << std::endl;
@@ -26,6 +26,6 @@ int main(int argc, char* argv[])
     webui::window win;
     win.set_root_folder(INDEX_PATH.string());
     win.show_browser("index.html", AnyBrowser);
-    win.bind("submit_url", handle_submit_url);
+    win.bind("submit_url", ytweb::handle_submit_url);
     webui::wait();
 }
